check if file exists
git-svn-id: https://svn.liip.ch/repos/public/fluxcms/trunk@10215 99f7f646-bdd5-0310-89f9-ddcb10b10bf3

+++ inc/bx/editors/blog.php
@@ -185,8 +185,6 @@ class bx_editors_blog extends bx_editor implements bxIeditor {
                 fwrite ($fd, '</categories>');
                 
                 fwrite($fd, '</entry>');
-		
-				
                 fclose($fd);
 				if ( "/".$data['uri'].".html" != $id) {
 					
@@ -194,10 +192,12 @@ class bx_editors_blog extends bx_editor implements bxIeditor {
 						// send ok...
 					} else if (empty($_POST['ajax'])  && empty($_GET['ajax'])) {
 						header("Location: ".$data['uri'] .".html");
-					} else {
-						unlink(BX_DATA_DIR."".$colluri."storage.xml");
-						header("Location: ".$data['uri'] .".html?ajax=1");
-					}
+					} else  {
+					    if (file_exists(BX_DATA_DIR."".$colluri."storage.xml")) {
+					        unlink(BX_DATA_DIR."".$colluri."storage.xml");
+					    }
+					    header("Location: ".$data['uri'] .".html?ajax=1");
+				    }
                 }
             }
         }
